[REFACTOR] useScrap 내부의 useMutation을 hooks/queries로 옮김
